Add estimateApproxFee function.
This is a more useful fee estimation function as if there is no fee high enough for the target you are aiming for, it will give you the fee estimate for the lowest target that you can reliably obtain.  This is better than defaulting to the minimum fee.  It will also pass back the target for which it returned an answer.